Only refactor
