+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/CodyAction.java
@@ -6,7 +6,6 @@ import com.intellij.notification.Notifications;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.DumbAwareAction;
-import com.sourcegraph.Icons;
 import org.jetbrains.annotations.NotNull;
 
 public class CodyAction extends DumbAwareAction {
