+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/completions/CompletionsInput.java
@@ -1,4 +1,4 @@
-package com.sourcegraph.completions;
+package com.sourcegraph.cody.completions;
 
 import org.jetbrains.annotations.NotNull;
