+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/completions/CompletionsService.java
@@ -1,4 +1,4 @@
-package com.sourcegraph.completions;
+package com.sourcegraph.cody.completions;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
