+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/completions/GraphQLWrapper.java
@@ -1,4 +1,4 @@
-package com.sourcegraph.completions;
+package com.sourcegraph.cody.completions;
 
 import com.google.gson.Gson;
 import org.jetbrains.annotations.NotNull;
