+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/completions/Speaker.java
@@ -1,4 +1,4 @@
-package com.sourcegraph.completions;
+package com.sourcegraph.cody.completions;
 
 public enum Speaker {
     HUMAN,
