+++ client/cody-jetbrains/src/main/java/com/sourcegraph/cody/editor/EditorContext.java
@@ -1,4 +1,4 @@
-package com.sourcegraph.cody;
+package com.sourcegraph.cody.editor;
 
 import org.jetbrains.annotations.Nullable;
